LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/scroll-timeline-in-container-query-expected.txt


FAIL Timeline appearing via container queries assert_equals: expected "rgb(0, 0, 0)" but got "rgb(100, 100, 100)"
